// bedes-common/models/xml-term/xml-term-option.interface.ts
export interface IXmlTermOption {
    uuid?: string | null | undefined;
    url?: string | null | undefined;
    listOption: string | null | undefined;
    listOptionDefinition?: string | null | undefined;
    unitName?: string | null | undefined;
    relatedTerm?: string | null | undefined;
    relatedTermUUID?: string | null | undefined;
    sectorNames?: Array<string> | null | undefined;
    applicationNames?: Array<string> | null | undefined;
    updatedDate?: string | null | undefined;
}